feat 213 : 백엔드 코드 수정, 마무리(#5)

+++ backend/src/main/java/com/mysite/knitly/domain/mypage/dto/OrderCardResponse.java
@@ -3,14 +3,15 @@ package com.mysite.knitly.domain.mypage.dto;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.math.BigDecimal;
 
 public record OrderCardResponse(
         Long orderId,
         LocalDateTime orderedAt,
-        Double totalPrice,
+        BigDecimal totalPrice,
         List<OrderLine> items
 ) {
-    public static OrderCardResponse of(Long orderId, LocalDateTime orderedAt, Double totalPrice) {
+    public static OrderCardResponse of(Long orderId, LocalDateTime orderedAt, BigDecimal totalPrice) {
         return new OrderCardResponse(orderId, orderedAt, totalPrice, new ArrayList<>());
     }
 }
