+++ backend/src/main/java/com/mysite/knitly/domain/mypage/dto/OrderLine.java
@@ -1,8 +1,10 @@
 package com.mysite.knitly.domain.mypage.dto;
 
+import java.math.BigDecimal;
+
 public record OrderLine(
         Long productId,
         String productTitle,
         int quantity,
-        Double orderPrice
+        BigDecimal orderPrice
 ) {}
